\name{lhs.dir.rhs}
\alias{lhs.dir.rhs}
\title{Separates from a vector of (in-)equalities the lhs, comparison operator and rhs}
\usage{
  lhs.dir.rhs(str)
}
\description{
  Separates from a vector of (in-)equalities the lhs,
  comparison operator and rhs
}

\examples{\dontrun{
  lhs.dir.rhs(str<-c("x+5>=y","a+4", "4=3"))
}}
