\name{maxima.connection}
\alias{maxima.connection}
\title{Interface between R and Maxima
Commands to establish connection to Maxima}
\usage{
  maxima.connection(out.path = NULL, maxima.path = NULL,
    use.pipe = TRUE, use.last = TRUE, use.log.file = TRUE)
}
\description{
  Interface between R and Maxima Commands to establish
  connection to Maxima
}

