\name{maxima.to.r}
\alias{maxima.to.r}
\title{Interface between R and Maxima
Calls Maxima from command line or opens a pipe
So far only tested under Windows}
\usage{
  maxima.to.r(str)
}
\description{
  Interface between R and Maxima Calls Maxima from command
  line or opens a pipe So far only tested under Windows
}

