\name{mx.check.concavity}
\alias{mx.check.concavity}
\title{Returns sufficient conditions from a principal minor test
that the function f is jointly concave in in all variables in var}
\usage{
  mx.check.concavity(f, var, assume = NULL)
}
\arguments{
  \item{f}{a string that specifies the Maxima formula of
  the function body}

  \item{var}{a character vector specifying the variables}
}
\description{
  Returns sufficient conditions from a principal minor test
  that the function f is jointly concave in in all
  variables in var
}

