\name{start.maxima}
\alias{start.maxima}
\title{Interface between R and Maxima
Commands to establish connection to Maxima}
\usage{
  start.maxima(out.path = tempdir(), maxima.path = NULL,
    use.pipe = TRUE, use.last = TRUE, use.log.file = TRUE)
}
\description{
  Interface between R and Maxima Commands to establish
  connection to Maxima
}

