\name{mx.quasi.concave.cond}
\alias{mx.quasi.concave.cond}
\title{Returns the conditions from the principal minor test
that the function f is jointly quasi-concave in in all variables in var}
\usage{
  mx.quasi.concave.cond(f, var, with.descr = FALSE,
    jacobian = NULL, hessian = NULL)
}
\arguments{
  \item{f}{a string that specifies the Maxima formula of
  the function body}

  \item{var}{a character vector specifying the variables}
}
\description{
  Returns the conditions from the principal minor test that
  the function f is jointly quasi-concave in in all
  variables in var
}

